+++ CHANGELOG.md
@@ -15,6 +15,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Slightly smaller and improved layout of content in case PDF report
 - Relabel more cancer variant pages somatic for navigation
 - Removed unused `add_compounds` param from variant controllers function
+- Only one ClinVar submission variant contition type per variant is used
+- ClinVar submission variant condition prefers OMIM over HPO if available
 
 ## [4.49]
 ### Fixed